130 - Fix MRCC clock input placement issue on Cmod A7 (XC7A15T-1CPG236C)
Vivado failed implementation because the SPI SCK pin was initially
assigned to a non-MRCC pin, leading to a clock routing violation.

Original error message:
[Place 30-574] Poor placement for routing between an IO pin and BUFG.
If this sub optimal condition is acceptable for this design, you may
use the CLOCK_DEDICATED_ROUTE constraint in the .xdc file to demote
this message to a WARNING. However, the use of this override is highly
discouraged.
<sck_IBUF_inst (IBUF.O) is locked to IOB_X0Y39
 sck_IBUF_BUFG_inst (BUFG.I) is provisionally placed by clockplacer on BUFGCTRL_X0Y0>

Root cause: clock net was assigned to a regular IO instead of an MRCC pin.
Fix: reassigned SCK input to a proper MRCC pin according to the Cmod A7 master XDC.